Fix console errors; refactor ingredientsList components
