+++ src/components/organisms/recipe/detail/ingredientsList/components/item.jsx
@@ -9,37 +9,31 @@ import {
 } from '@material-ui/core';
 import CommentIcon from '@material-ui/icons/Comment';
 
-import IconButton from '../../../../atoms/buttons/icon.jsx';
-import Checkbox from '../../../../atoms/fields/checkbox.jsx';
+import IconButton from '../../../../../atoms/buttons/icon.jsx';
+import Checkbox from '../../../../../atoms/fields/checkbox.jsx';
 
 const Component = ({
-  code,
   name,
   emoji,
   quantity,
   isOptional,
   alternatives,
-  key,
-  rowCounter,
-  checked,
-  handleToggle
+  value,
+  handleClick
 }) => {
   let label = emoji ? `${emoji} (${name})` : name;
   label = quantity ? `${quantity} ${label}` : label;
 
   return <ListItem
-    id={`${code}-${key}-${rowCounter}`}
-    key={`${code}-${key}-${rowCounter}`}
     role={undefined}
     dense
     button
-    onClick={handleToggle(rowCounter)}
+    onClick={handleClick}
   >
     <ListItemIcon>
-      <Checkbox value={checked.indexOf(rowCounter) !== -1} />
+      <Checkbox value={value} />
     </ListItemIcon>
     <ListItemText
-      id={`checkbox-list-label-${code}`}
       primary={isOptional ? `{${label}}` : `${label}`}
     />
     {alternatives && alternatives.length > 0 && (
@@ -53,16 +47,13 @@ const Component = ({
 };
 
 Component.propTypes = {
-  code: PropTypes.string,
   name: PropTypes.string,
   emoji: PropTypes.string,
   quantity: PropTypes.string,
-  isOptional: PropTypes.boolean,
+  isOptional: PropTypes.bool,
   alternatives: PropTypes.array,
-  key: PropTypes.number,
-  rowCounter: PropTypes.number,
-  checked: PropTypes.array,
-  handleToggle: PropTypes.Function,
+  value: PropTypes.bool,
+  handleToggle: PropTypes.any,
 };
 
 export default Component;
